Add mcu basics and chromiumcu
This is a work poc to provide the API for mcus to connect to our server
and provide additional non-peer to peer functionality.

As an example, this commit also adds chromium based javascript mcu which
can be used as a replacement of the Janus MCU videocall plugin. This
works as a drop-in replacement for Mattermost video calls.

References: KWM-22