Implement external TURN service support
Setting up a TURN server might be complicated as it requires public IPs
and bandwith. For this reason Kopano provides a simple TURN service
which can be used by paying Kopano customers to use a TURN server
provided by Kopano. This change adds the client and options to let KWM
server use a simple remote TURN service API with authentication.