Support replace of connections in groups
Whenever a new connection of a user comes in which is already in a
group, that connection gets rejected since the user is already active
with a connection. This change implements automatic replace of such
connections by help of a replace flag which is sent to the old
connection whenever a new connection comes in for a group which already
has a connection active for the same id.