Add automatic survey reporting
This change adds automatic reporting to Kopano for user survey. On
startup and henceforth every 60 minutes kwmserverd will send information
about itself to `stats.kopano.io` via HTTPS.

This behavior is turned on by default and can be turned off by setting
the environment variable `KOPANO_SURVEYCLIENT_AUTOSURVEY` to `false`.

With this commit, the following information is collected and reported:

  - Machine ID as found in /etc/machine-id
  - Program name (kwmserverd)
  - Compile time version of kwmserverd (kwmserverd --version)
  - Operating system name (uname -o)
  - Host machine hardware name (uname -m)
  - Kernel release (urname -r)
  - Pretty distribution name (either from /etc/os-release or
    /etc/redhat-release)